+++ src/pages/api/v1/cart.ts
@@ -1,0 +1,10 @@
+import { makeSecureHandler } from 'utils/proxy-handler';
+
+export const config = {
+  api: {
+    bodyParser: false,
+    externalResolver: true
+  }
+};
+
+export default makeSecureHandler(['GET', 'PUT', 'POST', 'DELETE']);
